backend/django/requirements.txt
asgiref==3.8.1
Django==5.1
psycopg2-binary==2.9.9
djangorestframework==3.15.2
sqlparse==0.5.1
